requirements.txt
Django==4.0.4
Pillow==9.1.0
djangorestframework==3.13.1
django-upload-validator==1.1.6